ファイル移動、一部修正
